+++ pytest_plt/tests/test_plt.py
@@ -42,3 +42,11 @@ def test_bbox_extra_artists(plt):
 def test_saveas(plt):
     assert plt.saveas.endswith("saveas.pdf")
     plt.saveas = None
+
+
+def test_saveas_pickle(plt):
+    axes = plt.subplots(2, 3)[1]  # The pickled figure will contain six axes.
+    x = np.linspace(-1, 1, 21)
+    for k, ax in enumerate(axes.ravel()):
+        ax.plot(x, x ** k)
+    plt.saveas = "%s.pickle" % (plt.saveas[:-4],)